updated basictree sample
